+++ src/05-chunker/api/index.js
@@ -3,6 +3,8 @@ import sentences from './sentences/index.js'
 import nouns from './nouns/index.js'
 import selections from './selections/index.js'
 import abbreviations from './abbreviations.js'
+import people from './people/index.js'
+import places from './places/index.js'
 import clauses from './clauses.js'
 
 const chunker = function (View) {
@@ -10,6 +12,8 @@ const chunker = function (View) {
   selections(View)
   verbs(View)
   sentences(View)
+  people(View)
+  places(View)
   abbreviations(View)
   clauses(View)
 }