+++ src/05-chunker/api/people/find.js
@@ -1,0 +1,5 @@
+const find = function (doc) {
+  let m = doc.match('#Person+')
+  return m
+}
+export default find